backend/chart_uploader_service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
